% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_coding.R
\name{DNA_To_inter_2mer}
\alias{DNA_To_inter_2mer}
\title{Encoding 2mer to transform a short DNA sequence into a data matrix containing only 0，1}
\usage{
DNA_To_inter_2mer(seq)
}
\description{
The input argument is a short DNA sequence
Output is a data matrix
@param seq a short DNA sequence
}
