% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_coding.R
\name{DNA_To_inter_onehot}
\alias{DNA_To_inter_onehot}
\title{Using onehot coding to transform DNA sequence into a data matrix containing only 0，1}
\usage{
DNA_To_inter_onehot(seq)
}
\arguments{
\item{seq}{a short DNA sequence}
}
\description{
The input argument is a short DNA sequence
Output is a data matrix
}
