% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methylation_DNA_coding.R
\name{get_all_onehot_2mer}
\alias{get_all_onehot_2mer}
\title{Extract specific length DNA sequence information of all CpG sites from cov file}
\usage{
get_all_onehot_2mer(
  flank_size = 50,
  ref_genome_dir = "ref_hg38/hg38/",
  CpG_report_file_name = "data/Methyl/H1_hESC.0.3.CpG_report.txt",
  pre_DNA_inter_dir = "data/All_hg38_DNA/"
)
}
\arguments{
\item{flank_size}{Side chain length, total window width 2*flank_size}

\item{ref_genome_dir}{Reference genome file directory, and individual chromosomes are stored separately in .fa file.}

\item{CpG_report_file_name}{The file contains "chr" and "position" in ".txt" file}

\item{pre_DNA_inter_dir}{The folder of output file}
}
\value{
Three ".Rdata" files in the pre_DNA_inter_dir.
}
\description{
Extract specific length DNA sequence information of all CpG sites from cov file
}
