% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNA_coding.R
\name{get_DNA_inter}
\alias{get_DNA_inter}
\title{Using the genomic location to get a short sequence then to transform a data matrix containing only 0，1 with single core computing}
\usage{
get_DNA_inter(
  center_chr_loci,
  flank_size = 100,
  fre_gen_dir = "data/hg38/",
  OnehotOr2mer = 1
)
}
\arguments{
\item{center_chr_loci}{dataframe or matrix that containing genomic changes and locations}

\item{flank_size}{Side chain length, total window width 2*flank_size}

\item{fre_gen_dir}{Reference genome file directory, and individual chromosomes are stored separately in .fa file.}

\item{OnehotOr2mer}{Index of 1 or 2，1 for onehot, 2 for 2mer}
}
\value{
A matrix contain 0,1
}
\description{
Using the genomic location to get a short sequence then to transform a data matrix containing only 0，1 with single core computing
}
\examples{
center_chr_loci<-data.frame(chr=c("chr1","chr2"),Position=c(100000,200000))
get_DNA_inter(center_chr_loci, 100, "data/hg38/", 1) 
}
