% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methylation_DNA_coding.R
\name{make_tain_Rdata}
\alias{make_tain_Rdata}
\title{Extract training samples from the coverage file after Bismark call methylation}
\usage{
make_tain_Rdata(
  cov_filename = "data/Methyl/ENCFF003FWN.0.5.deduplicated.bismark.cov.gz",
  flank_size = 100,
  output_prefix = "data/Train_Test_Data/ENCFF003FWN.0.5",
  train_size = 50000,
  ref_genome_dir,
  file_type = ".gz"
)
}
\arguments{
\item{cov_filename}{The file contain "chr", "start", "end", "methylation", "m.read" and "um.read"}

\item{flank_size}{Side chain length, total window width 2*flank_size}

\item{output_prefix}{prefix of output file}

\item{train_size}{Training sample size}

\item{ref_genome_dir}{Reference genome file directory, and individual chromosomes are stored separately in .fa file.}

\item{file_type}{cov_file type ".gz" or ".txt"}
}
\value{
A list containing Flank_methyl,Flank_DNA_onehot,Flank_DNA_2mer,Center_loci_table,Flank_m and Flank_um)
}
\description{
Extract training samples from the coverage file after Bismark call methylation
}
