% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_model_lowcov.R
\name{model_2mer_methyl_train}
\alias{model_2mer_methyl_train}
\title{Training the CNN model}
\usage{
model_2mer_methyl_train(
  train_data = "data/Train_Test_Data/GM12878.0.3.train.Rdata",
  figure_history_prefix = "pre_model2_2mer_methyl_GM12878",
  flank_size = 50,
  save_model = "model_2mer_methyl_GM12878.0.3.h5",
  epochs_Num = 20,
  batch_size_Num = 512,
  validation_split_Num = 0.2,
  poolingsize = 2
)
}
\arguments{
\item{train_data}{The train dataset produced by "make_tain_Rdata()". A list containing Flank_methyl,Flank_DNA_onehot,Flank_DNA_2mer,Center_loci_table,Flank_m and Flank_um)}

\item{figure_history_prefix}{The output of the history during training the model}

\item{flank_size}{Side chain length, total window width 2*flank_size}

\item{save_model}{Save the taining model}

\item{epochs_Num}{Set the CNN model's epochs}

\item{batch_size_Num}{Set the CNN model's batch_size}

\item{validation_split_Num}{Set the CNN model's validation_split}

\item{poolingsize}{Set the poolingsize. Recommended setting is 2}
}
\value{

}
\description{
Training the CNN model
}
