% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Method_model_lowcov.R
\name{Predict_test_low_chr}
\alias{Predict_test_low_chr}
\title{Prediction of low coverage sites with trained models.}
\usage{
Predict_test_low_chr(
  load_model_hdf5_file,
  testfile,
  flank_size,
  Predicted_file_name
)
}
\arguments{
\item{load_model_hdf5_file}{The taining model file path}

\item{testfile}{File path of the site to be predicted. The testfile produced by "make_test_Rdata()" or "make_lowcov_test_data_from_all". A list containing Flank_methyl,Flank_DNA_onehot,Flank_DNA_2mer,Center_loci_table,Flank_m and Flank_um)}

\item{flank_size}{Side chain length, total window width 2*flank_size}

\item{Predicted_file_name}{Output file after prediction. The data.frame contains "chr","start","m.read","um.read" and "estPredict". "estPredict" is the prediction value.}
}
\value{

}
\description{
Prediction of low coverage sites with trained models.
}
